+++ src/server/api/forum/getTags.ts
@@ -50,7 +50,6 @@ const getTags = {
             }
         ]);
 
-        console.log(tagUsage);
         return tagUsage;
     })
 };
